requirement.txt
streamlit>=1.20
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
PyPDF2
pandas
requests